Extract example parsers from interface.
